+++ packages/terra-menu/CHANGELOG.md
@@ -3,11 +3,16 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
 
 6.8.0 - (July 30, 2019)
 ------------------
 ### Removed
 * Removed DEPENDENCIES.md file
+
+### Changed
+* Fixed linter issue in doc
 
 6.7.1 - (July 23, 2019)
 ------------------