+++ packages/terra-menu/CHANGELOG.md
@@ -3,6 +3,26 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
+
+6.8.0 - (July 30, 2019)
+------------------
+### Removed
+* Removed DEPENDENCIES.md file
+
+### Changed
+* Fixed linter issue in doc
+
+6.7.1 - (July 23, 2019)
+------------------
+### Changed
+* Updated DEPENDENCIES.md doc
+
+6.7.0 - (July 23, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 6.6.0 - (July 16, 2019)
 ------------------